requirements.txt
numpy

opencv-python

PyAutoGUI

PyGetWindow

pyperclip

PyRect

PyScreeze

pytweening

pywin32

Pillow

orjson

loguru

httpx

tqdm

pynput

aiohttp

questionary

pyuac

requests

